Move event log and related components into separate folders and files
