+++ src/components/EventLog/EventDetails/DetailFields/RadiationDetailFields.tsx
