+++ src/components/EventLog/EventDetails/DetailFields/StemCellTransplantDetailFields.tsx
