+++ src/components/EventLog/EventDetails/DetailFields/SurgeryDetailFields.tsx
