+++ src/components/EventLog/EventDetails/DetailFields/prependDetailFieldName.ts
