+++ src/components/EventLog/EventDetails/DetailValues/DetailValue.tsx
