+++ src/components/EventLog/EventDetails/DetailValues/DetailValuesProps.ts
