+++ src/components/EventLog/EventDetails/DetailValues/DetailValuesWrapper.tsx
