+++ src/components/EventLog/EventDetails/DetailValues/EventTypes/StemCellTransplantDetails.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import DetailValue from "../DetailValue";
 import DetailValuesProps from "../DetailValuesProps";
 
-import { StemCellTransplantDetailValues } from "../../../../types/PatientEvent/Details/EventTypes/StemCellTransplantDetails";
+import { StemCellTransplantDetailValues } from "../../../../../types/PatientEvent/Details/EventTypes/StemCellTransplantDetails";
 
 export default function StemCellTransplantDetails(props: DetailValuesProps<StemCellTransplantDetailValues>) {
   return (
